ワーニングを除去した

+++ CGES/src/renderer_base.hpp
@@ -22,8 +22,8 @@ protected:
   virtual void ParallelDraw(const Camera& camera, 
                             RenderBuffer& renderTarget, 
                             const Scene& scene,
-                            const int loopMin,
-                            const int loopMax,
+                            const size_t loopMin,
+                            const size_t loopMax,
                             const glm::vec3& initialPos,
                             const glm::vec3& screenVerticalVec,
                             const glm::vec3& screenHorizontalVec) const = 0;
